Flytt knapperad til egen fil
